Modified structure of location and compilation UT tests (#890)
1) Created common unit/CMakeLists.txt for internal folders.
2) unit/kernels & unit/native_api is self-compiled now
3) Disabled l0 compilatiom folder. Have to disband this folder into independent essences,like: api, device_topology, kernels and etc.

+++ tests/unit/device_topology/ipc_serialization/ipc_test.hpp
@@ -136,7 +136,7 @@ TEST_F(ipc_handles_fixture, ipc_unix_server_handles_serialize) {
         auto got = rx_conn->receive_ipc_memory(data_to_recv, received_rank);
         ipc_handles.swap(got);
 
-        UT_ASSERT(received_rank == *other_pid, "Received rank is invalid");
+        UT_ASSERT(static_cast<int>(received_rank) == *other_pid, "Received rank is invalid");
         size_t index = 0;
         for (auto& recv_ipc_handle : ipc_handles) {
             std::shared_ptr<ccl_device> owner_device = recv_ipc_handle->get_owner().lock();
